Resolved conflicts with the remote branch
