+++ app/src/main/java/com/nitmz/morphosis/scoobydoo/ScoobyDooHomeActivity.java
@@ -196,8 +196,7 @@ public class ScoobyDooHomeActivity extends AppCompatActivity
             Intent intent = new Intent(this, ScoobyDooHomeActivity.class);
             startActivity(intent);
         } else if (id == R.id.nav_profile) {
-
-            replaceFragments(UserProfileFragment.class,false)
+            replaceFragments(UserProfileFragment.class,false);
         } else if (id == R.id.nav_leaderboard) {
             replaceFragments(LeaderboardFragment.class, false);
         } else if (id == R.id.nav_developer) {
