--- app/src/main/res/layout/fragment_profile.xml
@@ -1,58 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.nitmz.morphosis.scoobydoo.ProfileFragment">
-
-    <com.nitmz.morphosis.CircleImageView
-        android:id="@+id/profile_image_view"
-        android:layout_width="250dp"
-        android:layout_height="250dp"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true" />
-
-    <TextView
-        android:id="@+id/name_text_view"
-        android:layout_width="match_parent"
-        android:layout_height="55dp"
-        android:layout_below="@+id/profile_image_view"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="12dp" />
-
-    <android.support.v7.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/rank_card"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentStart="true">
-        <TextView
-            android:id="@+id/rank_text_view"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/profile_image_view"
-            android:layout_alignParentStart="true"
-            android:layout_marginTop="73dp" />
-    </android.support.v7.widget.CardView>
-
-    <android.support.v7.widget.CardView
-        android:id="@+id/score_card"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/rank_card"
-        android:layout_alignParentStart="true"
-        android:layout_marginBottom="15dp">
-
-        <TextView
-            android:id="@+id/score_text_view"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:layout_below="@+id/rank_text_view"
-            android:layout_marginTop="25dp" />
-    </android.support.v7.widget.CardView>
-
-</RelativeLayout>
